requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1

google-cloud-datastore==2.8.2
python-jose
flask-cors
six
python-dotenv
authlib